requirements.txt
PyYAML==3.12
tweepy==3.5.0
